This PR changes the following to speed up the unix implementation:
* ioctl() is avoided if possible
* write() is called from another thread

Yoshi's Island is accelerated as follows:
|    | before (a3c0c46) | after (this PR) |
|----|----|----|
| Raspberry Pi 3B | 50 fps 6/60 | 60 fps 30/60 |
| Ubuntu 18.04 | 10-30 fps 7/60 | 60 fps 60/60 |

Raspberry Pi 3B:
* Cortex-A53, 1.2 GHz, 4 cores
* Raspberry Pi OS (aarch64)
* osspd-alsa
* gcc 8.3.0

Ubuntu 18.04:
* i7-2820QM
* Ubuntu 18.04 LTS (x86_64, VM)
* osspd-alsa
* gcc 7.5.0